refactor: SessionAuthenticationInterceptor, TokenAuthenticationInterceptor 중복 코드 추상화
