+++ src/main/java/nextstep/subway/auth/ui/AuthenticationInterceptor.java
@@ -1,0 +1,51 @@
+package nextstep.subway.auth.ui;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import nextstep.subway.auth.domain.Authentication;
+import nextstep.subway.auth.domain.AuthenticationToken;
+import nextstep.subway.member.application.CustomUserDetailsService;
+import nextstep.subway.member.domain.LoginMember;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public abstract class AuthenticationInterceptor implements HandlerInterceptor {
+
+    private CustomUserDetailsService customUserDetailsService;
+    private ObjectMapper objectMapper = new ObjectMapper();
+    private AuthenticationConverter authenticationConverter;
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+        AuthenticationToken token = convert(request);
+        Authentication authentication = authenticate(token);
+        afterAuthentication(request, response, authentication);
+        return false;
+    }
+
+    public Authentication authenticate(AuthenticationToken token) {
+        String principal = token.getPrincipal();
+        LoginMember userDetails = customUserDetailsService.loadUserByUsername(principal);
+        checkAuthentication(userDetails, token);
+
+        return new Authentication(userDetails);
+    }
+
+    public AuthenticationToken convert(HttpServletRequest request) throws IOException {
+        return authenticationConverter.convert(request);
+    }
+
+    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
+        if (userDetails == null) {
+            throw new RuntimeException();
+        }
+
+        if (!userDetails.checkPassword(token.getCredentials())) {
+            throw new RuntimeException();
+        }
+    }
+
+    protected abstract void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException;
+}
